requirements-arm.txt
tensorflow-aarch64
keras
sklearn
numpy
opencv-python
matplotlib
imutils